pyproject.toml
[project]
name = "project-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.1",
    "litellm>=1.81.13",
]
